Update milanesa.html with image and description
The milanesa.html webpage now has an appropriate image of the dish.
Also, the webpage has an image of a Milanesa De Res after the header.
This new image resides in a newly created 'images' directory

After the image, a description for the dish was added with a h2 'Description' header.
The description for Milanesa has also been created as a paragraph.